Handle case when nfd and nfkd of a character differ.
So far we stored full decomposition of a character, together with the
information whether it is kanonical or not. Nevertheless, character can
have a non-kanonical decomposition, which itself is kanonically
decomposable.

We still store full decomposition, but not including kanonical
continuations if the first decomposition was not kanonical. For
non-kanonical decomposition we also store a flag indicating whether any
of the decomposition character is further kanonically decomposable.